Split out empty states
